add: add ReviewControllerPublic + endpoints + rating calculation

+++ mainapp/src/main/java/ru/practicum/ewm/event/Event.java
@@ -75,6 +75,5 @@ public class Event {
     private List<Review> reviews;
 
     @Column(name = "rating")
-    private Float rating;
-
+    private double rating;
 }
